Move master and slave to lib
