+++ lib/master/master_test.go
