+++ lib/slave/cracker_test.go
