+++ lib/slave/hasher/common.go
