+++ lib/slave/hasher/pbkdf2.go
